Move datastoreAvailability checker to OrganisationService
